requirements.txt
aiogram~=2.18
aioredis~=2.0
environs~=9.0
asyncpg~=0.26
pyyaml~=6.0